pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "ai-copilot-system"
version = "0.1.0"
description = "Local-first RAG, agents, and observability scaffold with a FastAPI backend and React console."
requires-python = ">=3.11"
dependencies = [
    "chromadb>=0.5.20",
    "fastapi>=0.115.0",
    "httpx>=0.27.0",
    "pydantic-settings>=2.6.0",
    "psycopg[binary]>=3.2.0",
    "pypdf>=5.1.0",
    "python-multipart>=0.0.17",
    "uvicorn[standard]>=0.32.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
]

[tool.pytest.ini_options]
pythonpath = ["backend"]
testpaths = ["backend/tests"]

[tool.setuptools]
package-dir = {"" = "backend"}

[tool.setuptools.packages.find]
where = ["backend"]
include = ["app*"]
